+++ tensorforce/core/optimizers/evolutionary.py
@@ -134,4 +134,4 @@ class Evolutionary(Optimizer):
 
         with tf.control_dependencies(control_inputs=(applied,)):
             # Trivial operation to enforce control dependency
-            return [util.identity_operation(x=delta, dtype='float') for delta in deltas]
+            return [util.identity_operation(x=delta) for delta in deltas]
