+++ tensorforce/core/optimizers/tf_optimizer.py
@@ -60,9 +60,7 @@ class TFOptimizer(Optimizer):
         # Force loss value to be calculated.
         with tf.control_dependencies(control_inputs=(loss,)):
             # Trivial operation to enforce control dependency
-            previous_variables = [
-                util.identity_operation(x=variable, dtype='float') for variable in variables
-            ]
+            previous_variables = [util.identity_operation(x=variable) for variable in variables]
 
         # The actual tensorflow minimize op.
         with tf.control_dependencies(control_inputs=previous_variables):
